+++ HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-0.2.0 (unreleased)
+0.2.0 (2014-02-26)
 ++++++++++++++++++
 
 * Bottle support.